+++ packages/ornate/src/tools/CircleTool.ts
@@ -16,7 +16,9 @@ export class CircleTool extends Tool implements ICogniteOrnateTool {
     this.ornateInstance.isDrawing = true;
 
     // If we're over an item with a group attachment, add it there instead.
-    const groupName = e.target.attrs?.attachedToGroup;
+    const groupName =
+      e.target.attrs?.attachedToGroup || e.target.attrs?.inGroup;
+
     this.group = this.ornateInstance.stage.findOne(
       `#${groupName}`
     ) as Konva.Group;