+++ packages/ornate/src/tools/CommentTool.ts
@@ -71,7 +71,8 @@ export class CommentTool extends Tool implements ICogniteOrnateTool {
     const { drawingLayer } = this.ornateInstance;
 
     // If we're over an item with a group attachment, add it there instead.
-    const groupName = e.target.attrs?.attachedToGroup;
+    const groupName =
+      e.target.attrs?.attachedToGroup || e.target.attrs?.inGroup;
     this.group = this.ornateInstance.stage.findOne(
       `#${groupName}`
     ) as Konva.Group;