+++ packages/ornate/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.3
+
+- BUG: Ensure documents are always visible on export
+- BUG: Use inGroup as well as attached group to determine if shape should be added to group
+- BUG: Ensure x and y is always set
+
 ## 1.1.2
 
 - BUG: Add high z index to text tool input field